Refactor FromContainer for reuse
